--- a/tests/queries/0_stateless/02286_parallel_final.reference
+++ b/tests/queries/0_stateless/02286_parallel_final.reference
@@ -1,9 +1,13 @@
+Test intersecting ranges
 2
 2
 3
 5
-8
-8
-8
-8
-8
+Test intersecting ranges finished
+Test non intersecting ranges
+0
+0
+0
+0
+0
+Test non intersecting ranges finished