--- a/tests/queries/0_stateless/02946_merge_tree_final_split_ranges_by_primary_key.sql
+++ b/tests/queries/0_stateless/02946_merge_tree_final_split_ranges_by_primary_key.sql
@@ -9,7 +9,7 @@ INSERT INTO test_table SELECT 0, '0';
 INSERT INTO test_table SELECT number + 1, number + 1 FROM numbers(15);
 OPTIMIZE TABLE test_table;
 
-SELECT COUNT() FROM system.parts WHERE table = 'test_table' AND active = 1;
+SELECT COUNT() FROM system.parts WHERE database = currentDatabase() AND table = 'test_table' AND active = 1;
 SYSTEM STOP MERGES test_table;
 
 SELECT '--';