… tests/queries/0_stateless/02875_final_invalid_read_ranges_bug.reference
… tests/queries/0_stateless/02875_final_invalid_read_ranges_bug.sql
… tests/queries/0_stateless/02946_merge_tree_final_split_ranges_by_primary_key.reference