@kitaisreal 2023-12-21
Table without duplicates:

```
CREATE TABLE test_table (id UInt64, value String) ENGINE=ReplacingMergeTree ORDER BY id SETTINGS index_granularity = 8192;
INSERT INTO test_table SELECT number, number FROM numbers(500000000);

OPTIMIZE TABLE test_table;

SELECT name, part_type, marks FROM system.parts WHERE table = 'test_table' AND active = 1;

┌─name────────┬─part_type─┬─marks─┐
│ all_1_477_4 │ Wide      │ 61037 │
└─────────────┴───────────┴───────┘
```

Before:
```
SELECT * FROM test_table FINAL FORMAT Null

0 rows in set. Elapsed: 3.464 sec. Processed 500.12 million rows, 12.89 GB (144.36 million rows/s., 3.72 GB/s.)
Peak memory usage: 178.61 MiB.
```

After:
```
SELECT * FROM test_table FINAL FORMAT Null

0 rows in set. Elapsed: 0.525 sec. Processed 500.00 million rows, 12.89 GB (952.04 million rows/s., 24.54 GB/s.)
Peak memory usage: 43.50 MiB.
```

The same as without final:
```
SELECT * FROM test_table FORMAT Null

0 rows in set. Elapsed: 0.526 sec. Processed 500.00 million rows, 12.89 GB (950.44 million rows/s., 24.50 GB/s.)
Peak memory usage: 40.77 MiB.
```

Table with some intersecting ranges:

```
DROP TABLE IF EXISTS test_table_intersecting_ranges;
CREATE TABLE test_table_intersecting_ranges (id UInt64, value String) ENGINE=ReplacingMergeTree ORDER BY id SETTINGS index_granularity = 8192;
INSERT INTO test_table_intersecting_ranges SELECT number, number FROM numbers(250000000);
INSERT INTO test_table_intersecting_ranges SELECT number + 200000000, number FROM numbers(250000000);
SYSTEM STOP MERGES test_table_intersecting_ranges;
```

Before:
```
SELECT * FROM test_table_intersecting_ranges FINAL FORMAT Null

0 rows in set. Elapsed: 3.279 sec. Processed 500.16 million rows, 12.78 GB (152.53 million rows/s., 3.90 GB/s.)
Peak memory usage: 196.35 MiB.
```

After:
```
SELECT * FROM test_table_intersecting_ranges FINAL FORMAT Null

0 rows in set. Elapsed: 1.379 sec. Processed 500.06 million rows, 12.78 GB (362.61 million rows/s., 9.27 GB/s.)
Peak memory usage: 128.50 MiB.
```

Setting do_not_merge_across_partitions_select_final performance improvements:

Create table:
```
DROP TABLE IF EXISTS test_table_partitioned;
CREATE TABLE test_table_partitioned (id UInt64, value String) ENGINE=ReplacingMergeTree ORDER BY id PARTITION BY id % 10 SETTINGS index_granularity = 8192;
INSERT INTO test_table_partitioned SELECT number, number FROM numbers(500000000);
OPTIMIZE TABLE test_table_partitioned;

SELECT name, part_type, marks FROM system.parts WHERE table = 'test_table_partitioned' AND active = 1;

┌─name────────┬─part_type─┬─marks─┐
│ 0_1_4767_5  │ Compact   │  6105 │
│ 1_2_4768_5  │ Compact   │  6105 │
│ 2_3_4769_5  │ Compact   │  6105 │
│ 3_4_4770_5  │ Compact   │  6105 │
│ 4_5_4761_5  │ Compact   │  6105 │
│ 5_6_4762_5  │ Compact   │  6105 │
│ 6_7_4763_5  │ Compact   │  6105 │
│ 7_8_4764_5  │ Compact   │  6105 │
│ 8_9_4765_5  │ Compact   │  6105 │
│ 9_10_4766_5 │ Compact   │  6105 │
└─────────────┴───────────┴───────┘
```

Before:

```
SELECT * FROM test_table_partitioned FINAL SETTINGS do_not_merge_across_partitions_select_final = 1 FORMAT Null

0 rows in set. Elapsed: 0.891 sec. Processed 500.00 million rows, 12.89 GB (560.99 million rows/s., 14.46 GB/s.)
Peak memory usage: 32.01 MiB.
```

After:

```
SELECT * FROM test_table_partitioned FINAL SETTINGS do_not_merge_across_partitions_select_final = 1 FORMAT Null

0 rows in set. Elapsed: 0.610 sec. Processed 500.00 million rows, 12.89 GB (819.56 million rows/s., 21.13 GB/s.)
Peak memory usage: 54.96 MiB.
```