@kitaisreal 2023-12-21
> For ReplacingMergeTree, test like that is more likely
> 
> ```
> DROP TABLE IF EXISTS test_table_intersecting_ranges;
> CREATE TABLE test_table_intersecting_ranges (id UInt64, value String) ENGINE=ReplacingMergeTree ORDER BY id SETTINGS index_granularity = 8192;
> INSERT INTO test_table_intersecting_ranges SELECT number , number FROM numbers(500000000);
> INSERT INTO test_table_intersecting_ranges SELECT number*50000, number FROM numbers(20);
> 
> SYSTEM STOP MERGES test_table_intersecting_ranges;
> ```
Thanks for test.

Before:
```
SELECT * FROM test_table_intersecting_ranges FINAL FORMAT Null

0 rows in set. Elapsed: 3.343 sec. Processed 500.12 million rows, 12.89 GB (149.62 million rows/s., 3.86 GB/s.)
Peak memory usage: 239.85 MiB.
```

After:
```
SELECT * FROM test_table_intersecting_ranges FINAL FORMAT Null

0 rows in set. Elapsed: 0.629 sec. Processed 489.72 million rows, 12.62 GB (779.03 million rows/s., 20.08 GB/s.)
Peak memory usage: 46.30 MiB.
```