@kitaisreal 2023-12-24
Performance results.

X86-64:
https://s3.amazonaws.com/clickhouse-test-reports/58120/a85e2cd6708ca3f6d6ce97e2243db85848b8047e/performance_comparison_[1_4]/report.html
https://s3.amazonaws.com/clickhouse-test-reports/58120/a85e2cd6708ca3f6d6ce97e2243db85848b8047e/performance_comparison_[4_4]/report.html

ARM:
https://s3.amazonaws.com/clickhouse-test-reports/58120/a85e2cd6708ca3f6d6ce97e2243db85848b8047e/performance_comparison_aarch64_[1_4]/report.html
https://s3.amazonaws.com/clickhouse-test-reports/58120/a85e2cd6708ca3f6d6ce97e2243db85848b8047e/performance_comparison_aarch64_[4_4]/report.html