@tavplubix 2023-12-28
The code is quite tricky and it handles many corner cases, so please consider introducing a setting like `optimize_non_intersecting_ranges_on_final` that allows disabling this optimization (just in case).

Also, there's a lack of introspection/logging, so it will be hard to debug potential issues